[SPARK-15899][SQL] Fix the construction of the file path with hadoop Path for Spark 2.0
This PR contains the adaptation of https://github.com/apache/spark/pull/13868 for Spark 2.0

## What changes were proposed in this pull request?

Fix the construction of the file path in `SQLConf.scala` and unit tests that rely on this: `SQLConfSuite` and `DDLSuite`. Previous way of construction caused the creation of incorrect path on Windows.

## How was this patch tested?

Run unit tests on Windows

Author: avulanov <nashb@yandex.ru>

Closes #14600 from avulanov/SPARK-15899-file-2.0.