[SPARK-16711] YarnShuffleService doesn't re-init properly on YARN rolling upgrade
branch-2.0 version of this patch.  The differences are in the YarnShuffleService for finding the location to put the DB. branch-2.0 does not use the yarn nm recovery path like master does.

Tested in manually on 8 node yarn cluster and ran unit tests.  Manually tests verified DB created properly and it found them if already existed. Verified that during rolling upgrade credentials were reloaded and running application was not affected.

Author: Thomas Graves <tgraves@staydecay.corp.gq1.yahoo.com>

Closes #14997 from tgravescs/SPARK-16711-branch2.0.