[SPARK-20197][SPARKR][BRANCH-2.1] CRAN check fail with package installation
## What changes were proposed in this pull request?

Test failed because SPARK_HOME is not set before Spark is installed.
Also current directory is not == SPARK_HOME when tests are run with R CMD check, unlike in Jenkins, so disable that test for now. (that would also disable the test in Jenkins - so this change should not be ported to master as-is.)

## How was this patch tested?

Manual run R CMD check

Author: Felix Cheung <felixcheung_m@hotmail.com>

Closes #17515 from felixcheung/rcrancheck.