Revert part of codes [BUG-79168] by removing username setting in the HiveSessionStateBuilder
The username from the system properties would be used in the llap. The HiveSessionStateBuilder get the system user name and stored it in the HiveSessionStateBuilder. However, the HiveSessionStateBuilder is not a session, it is not suitable to store this username there. We need to reimplement this for the spark 2.2. Thus, we just remove the current code.

Change-Id: I237aeeb9597eda85d301ee58ec82bc617e437cf3