[SPARK-20763][SQL] The function of `month` and `day` return the value which is not we expected.
## What changes were proposed in this pull request?
spark-sql>select month("1582-09-28");
spark-sql>10
For this case, the expected result is 9, but it is 10.

spark-sql>select day("1582-04-18");
spark-sql>28
For this case, the expected result is 18, but it is 28.

when the date  before "1582-10-04", the function of `month` and `day` return the value which is not we expected.

## How was this patch tested?
unit tests

Author: liuxian <liu.xian3@zte.com.cn>

Closes #17997 from 10110346/wip_lx_0516.

(cherry picked from commit ea3b1e352a605cd35cdee987d0e5eb8528ef1b45)
Signed-off-by: Xiao Li <gatorsmile@gmail.com>