[SPARK-24014][PYSPARK] Add onStreamingStarted method to StreamingListener
The `StreamingListener` in PySpark side seems to be lack of `onStreamingStarted` method. This patch adds it and a test for it.

This patch also includes a trivial doc improvement for `createDirectStream`.

Original PR is #21057.

Added test.

Author: Liang-Chi Hsieh <viirya@gmail.com>

Closes #21098 from viirya/SPARK-24014.

(cherry picked from commit 8bb0df2c65355dfdcd28e362ff661c6c7ebc99c0)
Signed-off-by: jerryshao <sshao@hortonworks.com>

Change-Id: I8dd3ef36aff0b96ff9a6b78878885a7c9e40d0d4