[SPARK-24309][CORE] AsyncEventQueue should stop on interrupt.
EventListeners can interrupt the event queue thread.  In particular,
when the EventLoggingListener writes to hdfs, hdfs can interrupt the
thread.  When there is an interrupt, the queue should be removed and stop
accepting any more events.  Before this change, the queue would continue
to take more events (till it was full), and then would not stop when the
application was complete because the PoisonPill couldn't be added.

Added a unit test which failed before this change.

Author: Imran Rashid <irashid@cloudera.com>

Closes #21356 from squito/SPARK-24309.

(cherry picked from commit 32447079e9d0fa9f7e180b94ecac19091b6af1ab)
Signed-off-by: Marcelo Vanzin <vanzin@cloudera.com>

Change-Id: I086f098815ed35a35e2c97042c3445f6ebfad830