[MINOR] Add port SSL config in toString and scaladoc
SPARK-17874 introduced a new configuration to set the port where SSL services bind to. We missed to update the scaladoc and the `toString` method, though. The PR adds it in the missing places

checked the `toString` output in the logs

Author: Marco Gaido <marcogaido91@gmail.com>

Closes #21429 from mgaido91/minor_ssl.

(cherry picked from commit fd315f5884c03c6dd21abca178897584dee83f1a)
Signed-off-by: Marcelo Vanzin <vanzin@cloudera.com>

Change-Id: I70ac42339a619ba44b670be891f8ee9aa3d446f7