[BUG-119744][SPARK-24495][SQL] EnsureRequirement returns wrong plan when reordering equal keys
`EnsureRequirement` in its `reorder` method currently assumes that the same key appears only once in the join condition. This of course might not be the case, and when it is not satisfied, it returns a wrong plan which produces a wrong result of the query.

added UT

Author: Marco Gaido <marcogaido91@gmail.com>

Closes #21529 from mgaido91/SPARK-24495.

(cherry picked from commit fdadc4be08dcf1a06383bbb05e53540da2092c63)
Signed-off-by: Xiao Li <gatorsmile@gmail.com>

Change-Id: Ib698cedbdd254487c391a5eec60a64f0c546b94c