[SPARK-25371][SQL][BACKPORT-2.3] struct() should allow being called with 0 args
SPARK-21281 introduced a check for the inputs of `CreateStructLike` to be non-empty. This means that `struct()`, which was previously considered valid, now throws an Exception.  This behavior change was introduced in 2.3.0. The change may break users' application on upgrade and it causes `VectorAssembler` to fail when an empty `inputCols` is defined.

The PR removes the added check making `struct()` valid again.

added UT

Closes #22391 from mgaido91/SPARK-25371_2.3.

Authored-by: Marco Gaido <marcogaido91@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
Change-Id: I81f82b6180ef96b886319e380164f664549f28a6